Extract the app logic to a separated file.
- Split the app setup logic to a single file.
- Easier to test app.
